Break down withKeys class
