+++ docs/openapi-docs/src/main/scala/tapir/docs/openapi/EndpointToOpenApiComponents.scala
@@ -1,0 +1,15 @@
+package tapir.docs.openapi
+import tapir.docs.openapi.ObjectSchemasForEndpoints.SchemaKeys
+import tapir.openapi.Components
+
+private[openapi] class EndpointToOpenApiComponents(schemaKeys: SchemaKeys) {
+
+  def components: Option[Components] = {
+    if (schemaKeys.nonEmpty) {
+      Some(Components(Some(schemaKeys.map {
+        case (_, (key, schema)) =>
+          key -> Right(schema)
+      })))
+    } else None
+  }
+}
